BOOO

+++ Content/Scripts/Singletons/UnchangableMeta.cs
@@ -16,11 +16,11 @@ public partial class UnchangableMeta : Node
     //the values of the LevelsCompleteStatus array variables can be 0, 1, 2, 3. 0 means the level is not passed, 1 means it is passed at the minimum difficulty,
     //correspondingly 2 at the average and 3 at the maximum difficulty.
     public static int[] LevelCompleteStatus = new int[G.LevelsInGameTotal];
-    public static byte[] LevelPlayedStatus = new byte[G.LevelsInGameTotal]; //I made it as byte[] because of retard Godot that can't save a boolean array >:(
+    public static int[] LevelPlayedStatus = new int[G.LevelsInGameTotal]; //I made it as byte[] because of retard Godot that can't save a boolean array >:( // UPD: On top of that, the JSON.Stringify method converts an array of bytes into a string. To avoid doing bdsm, I converted the bytes to int. Sorry.
 
     public static bool IsLanguageSetted = false, IsTutorialPlayed, IsLevel9PlatformSectionFirstTimeCompleted, IsLevel9PlatformSectionSkipAllowed, IsFakeLevel10SkipAllowed, IsThereNewContentInRecycleBin = true;
     public static bool DidModsCrushedTheGame = false;
-    public static byte[] HintsStatus = //1 - was showed. 0 - hasn't.
+    public static int[] HintsStatus = //1 - was showed. 0 - hasn't.
     {
         0, // id 0 is Level 2 standing penalty hint
     };
